Separate internal and public API GameLibrary code
The internal GameLibrary class does not need to be known by API users
and is only useful for interal API usage.
