+++ SlashGaming-Diablo-II-API/src/game_address.cc
@@ -48,7 +48,7 @@
 #include <fmt/printf.h>
 #include "../include/game_address_locator.h"
 #include "../include/default_game_library.h"
-#include "game_library_table.h"
+#include "game_library.h"
 #include "../include/game_version.h"
 
 #include "c/game_address.h"
