+++ SlashGaming-Diablo-II-API/src/game_address_table_reader.cc
@@ -53,7 +53,7 @@
 #include "../include/game_address.h"
 #include "../include/game_address_locator.h"
 #include "../include/default_game_library.h"
-#include "game_library_table.h"
+#include "game_library.h"
 #include "../include/game_version.h"
 
 namespace sgd2mapi {
