--- SlashGaming-Diablo-II-API/src/game_library_table.cc
@@ -1,115 +1,0 @@
-/**
- * SlashGaming Diablo II Modding API
- * Copyright (C) 2018-2019  Mir Drualga
- *
- * This file is part of SlashGaming Diablo II Modding API.
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as published
- *  by the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- *  Additional permissions under GNU Affero General Public License version 3
- *  section 7
- *
- *  If you modify this Program, or any covered work, by linking or combining
- *  it with Diablo II (or a modified version of that game and its
- *  libraries), containing parts covered by the terms of Blizzard End User
- *  License Agreement, the licensors of this Program grant you additional
- *  permission to convey the resulting work. This additional permission is
- *  also extended to any combination of expansions, mods, and remasters of
- *  the game.
- *
- *  If you modify this Program, or any covered work, by linking or combining
- *  it with any Glide wrapper (or a modified version of that library),
- *  containing parts not covered by a compatible license, the licensors of
- *  this Program grant you additional permission to convey the resulting
- *  work.
- */
-
-#include "game_library_table.h"
-
-#include <windows.h>
-#include <filesystem>
-#include <stdexcept>
-#include <string>
-#include <string_view>
-
-#include <fmt/format.h>
-#include <fmt/printf.h>
-#include "../include/default_game_library.h"
-
-namespace sgd2mapi {
-namespace {
-
-std::map<std::filesystem::path, GameLibrary>&
-GetLibrariesByPaths(
-    void
-) {
-  static std::map<std::filesystem::path, GameLibrary> libraries_by_paths;
-  return libraries_by_paths;
-}
-
-} // namespace
-
-
-const GameLibrary&
-GetGameLibrary(
-    const std::filesystem::path& library_path
-) {
-  std::map<std::filesystem::path, GameLibrary>& libraries_by_paths =
-      GetLibrariesByPaths();
-
-  if (libraries_by_paths.count(library_path) == 0) {
-    libraries_by_paths.emplace(
-        library_path,
-        GameLibrary(library_path)
-    );
-  }
-
-  try {
-    return GetLibrariesByPaths().at(library_path);
-  } catch (const std::out_of_range& e) {
-    constexpr std::wstring_view kErrorFormatMessage =
-        L"File: %s \n"
-        L"Line: %d \n"
-        L"Could not determine the game library from the file path: %s.";
-
-    std::wstring full_message = fmt::sprintf(
-        kErrorFormatMessage,
-        fmt::to_wstring(__FILE__),
-        __LINE__,
-        library_path
-    );
-
-    MessageBoxW(
-        nullptr,
-        full_message.data(),
-        L"Failed to Determine Game Library",
-        MB_OK | MB_ICONERROR
-    );
-
-    std::exit(0);
-  }
-}
-
-const GameLibrary&
-GetGameLibrary(
-    enum DefaultLibrary library
-) {
-  const std::filesystem::path& library_path =
-      GetDefaultLibraryPathWithRedirect(
-          library
-      );
-  return GetGameLibrary(library_path);
-}
-
-} // namespace sgd2mapi
